feat: Money store its value as BigInteger type

+++ src/main/java/lotto/domain/Money.java
@@ -1,21 +1,23 @@
 package lotto.domain;
 
+import java.math.BigInteger;
+
 public class Money {
     private static final int MIN_AMOUNT = 0;
-    private final int money;
+    private final BigInteger money;
 
-    private Money(int money) {
+    private Money(long money) {
         validateAmount(money);
-        this.money = money;
+        this.money = BigInteger.valueOf(money);
     }
 
-    private void validateAmount(int money) {
+    private void validateAmount(long money) {
         if (money < MIN_AMOUNT) {
             throw new MoneyLowerThanZeroException();
         }
     }
 
-    public static Money from(int money) {
+    public static Money from(long money) {
         return new Money(money);
     }
 }
